script refactoring
